requirements.txt
numpy
torchvision
matplotlib
Pillow
scipy
huggingface_hub
transformers
datasets
accelerate
scikit-learn